refactor: Validating ModalPacking
